pyproject.toml
[tool.poetry]
name = "base-django-template"
version = "0.1.0"
description = ""
authors = ["Tarlan Soltanov <tarlansoltanov@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

django = "^4.2.1"
django-redis = "^5.2.0"
django-health-check = "^3.17.0"
django-split-settings = "^1.2.0"

psycopg2 = "^2.9.6"
python-decouple = "^3.8"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1.0"
django-querycount = "^0.8.3"
nplusone = "^1.0.0"

flake8 = "^6.0.0"
flake8-django = "^1.3"

pytest = "^7.3.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"

dotenv-linter = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
